requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
click>=8.1.7
rich>=13.7.0
jinja2>=3.1.2
python-dateutil>=2.8.2
tenacity>=8.2.3
ratelimit>=2.2.1
pathlib>=1.0.1 